+++ dwds/lib/src/debugging/inspector.dart
@@ -448,7 +448,7 @@ class AppInspector implements AppInspectorInterface {
     if (result == null) {
       throw RPCError(
         'getMemoryUsage',
-        RPCError.kInternalError,
+        RPCErrorKind.kInternalError.code,
         'Null result from chrome Devtools.',
       );
     }
@@ -461,7 +461,7 @@ class AppInspector implements AppInspectorInterface {
     if (usage == null) {
       throw RPCError(
         'getMemoryUsage',
-        RPCError.kInternalError,
+        RPCErrorKind.kInternalError.code,
         'Failed to parse memory usage result.',
       );
     }
@@ -536,7 +536,7 @@ class AppInspector implements AppInspectorInterface {
     // breakpoints. This is because the token positions are derived from the
     // DDC source maps which Chrome also uses.
     final tokenPositions = <int>[
-      for (var location in mappedLocations) location.tokenPos
+      for (var location in mappedLocations) location.tokenPos,
     ];
     tokenPositions.sort();
 
@@ -735,7 +735,7 @@ class AppInspector implements AppInspectorInterface {
         final parts = scripts[uri];
         final scriptRefs = [
           ScriptRef(uri: uri, id: createId()),
-          for (var part in parts ?? []) ScriptRef(uri: part, id: createId())
+          for (var part in parts ?? []) ScriptRef(uri: part, id: createId()),
         ];
         final libraryRef = await _libraryHelper.libraryRefFor(uri);
         final libraryId = libraryRef?.id;