+++ dwds/test/expression_evaluator_test.dart
@@ -121,7 +121,7 @@ void main() async {
             'params': {
               'reason': 'other',
               'callFrames': [],
-            }
+            },
           }),
         );