+++ dwds/test/fixtures/context.dart
@@ -372,9 +372,9 @@ class TestContext {
                 'remote-debugging-port=$debugPort',
                 if (enableDebugExtension)
                   '--load-extension=debug_extension/prod_build',
-                if (headless) '--headless'
-              ]
-            }
+                if (headless) '--headless',
+              ],
+            },
           });
         _webDriver = await createDriver(
           spec: WebDriverSpec.JsonWire,